Add Player class
The Player class represents a player of the game, and can be both
human or non-human.

Each Player object holds a name, an instance of Gameboard, and whether
it is currently their turn.

startTurn() and endTurn() are used to toggle the player's turn.

attack() and attackRandom() allow a player to attack another specified
player.